@@ -13,13 +13,13 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/common": "~13.4.0",
-    "@angular/compiler": "~13.4.0",
-    "@angular/core": "~13.4.0",
-    "@angular/forms": "~13.4.0",
-    "@angular/platform-browser": "~13.4.0",
-    "@angular/platform-browser-dynamic": "~13.4.0",
-    "@angular/router": "~13.4.0",
+    "@angular/common": "^14.3.0",
+    "@angular/compiler": "^14.3.0",
+    "@angular/core": "^14.3.0",
+    "@angular/forms": "^14.3.0",
+    "@angular/platform-browser": "^14.3.0",
+    "@angular/platform-browser-dynamic": "^14.3.0",
+    "@angular/router": "^14.3.0",
     "@awesome-cordova-plugins/core": "^6.6.0",
     "@awesome-cordova-plugins/native-audio": "^6.6.0",
     "@awesome-cordova-plugins/splash-screen": "^6.6.0",
@@ -32,25 +32,25 @@
     "@capacitor/ios": "3.4.3",
     "@capacitor/keyboard": "1.0.2",
     "@capacitor/status-bar": "1.0.2",
-    "@ionic/angular": "^6.7.5",
-    "@ionic/storage-angular": "^3.0.6",
-    "@ngrx/store": "^12.4.0",
+    "@ionic/angular": "^7.8.4",
+    "@ionic/storage-angular": "^4.0.0",
+    "@ngrx/store": "^14.3.3",
     "cordova-plugin-nativeaudio": "^3.0.9",
     "moment": "2.30.1",
     "ngx-countdown": "^12.0.1",
-    "rxjs": "~6.6.0",
+    "rxjs": "^7.5.0",
     "tslib": "^2.2.0",
     "zone.js": "~0.11.4"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "~13.3.11",
-    "@angular-eslint/builder": "~12.0.0",
-    "@angular-eslint/eslint-plugin": "~12.0.0",
-    "@angular-eslint/eslint-plugin-template": "~12.0.0",
-    "@angular-eslint/template-parser": "~12.0.0",
-    "@angular/cli": "~13.3.11",
-    "@angular/compiler-cli": "~13.4.0",
-    "@angular/language-service": "~13.4.0",
+    "@angular-devkit/build-angular": "^14.2.13",
+    "@angular-eslint/builder": "~14.4.0",
+    "@angular-eslint/eslint-plugin": "~14.4.0",
+    "@angular-eslint/eslint-plugin-template": "~14.4.0",
+    "@angular-eslint/template-parser": "~14.4.0",
+    "@angular/cli": "^14.2.13",
+    "@angular/compiler-cli": "^14.3.0",
+    "@angular/language-service": "^14.3.0",
     "@capacitor/cli": "3.1.2",
     "@ionic/angular-toolkit": "^4.0.0",
     "@types/jasmine": "~3.6.0",
